requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
cffi==1.16.0
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
pycparser==2.21
PyNaCl==1.5.0
yarl==1.9.4
